COMPILED PORTFOLIO
This portfolio contains the analysis of the previous amazon top 50 bestsellers, as well as the analysis required, having been provided with data on courses from different topics on a particular learning platform, to understand where opportunities to increase revenue may lie. A presentation of the data on course revenue, and then tracking the performance of courses is required.